Separated out functions, improved releases data format
